[refactor] - 학교 상세 조회 usecase 연결 및 활동 이력 출력 개선(SchoolInfo 활용)

+++ data/src/main/java/com/umc/data/dataSource/remote/OrganizationRemoteDataSourceImpl.kt
@@ -47,7 +47,7 @@ class OrganizationRemoteDataSourceImpl @Inject constructor(
     override suspend fun getStudyGroupDetail(groupId: Long): ApiState<StudyGroupDetailResponse> =
         apiCall { organizationApi.getStudyGroupDetail(groupId) }
 
-    override suspend fun getSchoolDetail(schoolId: Int): ApiState<SchoolDetailResponse> {
+    override suspend fun getSchoolDetail(schoolId: Long): ApiState<SchoolDetailResponse> {
         return apiCall { organizationApi.getSchoolDetail(schoolId) }
     }
